add flags support of shmget syscall
